Refactor material upload and dashboard routers
重新組織上傳/儀錶板流程，並新增上傳服務和驗證。

主要變更：

- 引入了一個專用服務 (app/services/uploadmaterial.py)，用於保存帶有 UUID 文件名的上傳文件，並簡化了文字提取邏輯。

- 將上傳處理移至 app/routers/dashboard.py（現在以 /materials 為前綴），並新增了 POST /upload：檔案副檔名白名單、呼叫 save_upload_file、建立 UserDocument 資料庫條目並傳回上傳結果；包含錯誤處理。

- 將先前的 uploadmaterial 路由轉換為 app/routers/uploadmaterial.py，以便在 GET /me 請求中提供儀錶板資訊（greeting、profile、active_doc、stats），並將問候邏輯移至此處。

- 更新了 app/models/dashboard.py（少量格式調整），並在路由中新增了匯入/資料庫互動。

- 更新了 .gitignore 文件，忽略 studydy.db 和 uploads/ 目錄；從倉庫中移除了追蹤資料庫和許多已上傳的檔案。

- 新增了 tests/test_upload.py 文件，其中包含基本的上傳測試（無效擴展名和基本的 PDF 上傳情況）。

此次重構分離了關注點，強制執行上傳驗證，避免了檔案名稱衝突，並透過忽略/刪除產生的資源來清理程式碼庫。
