+++ app/models/dashboard.py
@@ -5,8 +5,8 @@ from typing import Optional
 class UserDocument(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     user_id: int = Field(foreign_key="user.id")
-    title: str
-    file_path: str
+    title: str 
+    file_path: str  
     progress: int = Field(default=0)
     last_accessed: datetime = Field(default_factory=datetime.utcnow)
